fix: switch from OCM token file to env var
The OCM UI now supplies a single access token, rather than a JSON
object. Therefore, rather than instructing users to write the JSON to a
file in /tmp, instruct them to set the environment variable OCM_TOKEN
and read from that.

Since we no longer need to parse JSON, the entire Token struct is no
longer needed and has been removed.

OCM_TOKEN was picked as the environment variable name to match a subset
of the behavior of the ocm and rosa CLIs, without attempting to
comprehensively implement their behavior, since in production, the
operator will get its configuration elsewhere.