chore: replace make vet and fmt with golangci-lint
In CI, this project already uses golangci-lint to configure and run all
of its various linters. Run the same thing locally in "make lint" and as
part of make targets that previously ran "go vet" and "go fmt". This has
the added benefit of commands like "make build" and "make test" causing
diffs, since it is difficut to run "go fmt" without it actually
modifying files.

Add a new "golangci-lint" target that downloads a project-specific copy
of golangci-lint to the project's bin directory, similar to what we do
for kustomize and other tools. The version is set to v1.52.2, which is
the same we have configured in CI.

Add a new "make lint" target to run the linter and get rid of the old
"vet" and "fmt" targets since our lint list checks those. Call this
target from everywhere that previously called "vet" and "fmt".