fix: run tests explicitly rather than in builds
Avoid the situation where our CI runs lint as part of its own action,
and then again implicitly as part of running builds or unit tests. CI
will no only run lint in its own action.

Remove the "test" make target as a dependency of the "docker-build"
target, and the "lint" make target as a dependency of the "build" target.
Instead, CI now explicitly calls a new target "make test-nolint" when
running unit tests. This target runs everything that "make test" used to
run except for linting. "make test" still exists for command-line usage.

Note that the "build" make target is not explicitly called in CI, but it
no longer depends upon "lint" so that it acts like the other targets
that perform builds.